CardExportable recibe seriesApi por parametro
    refs:#371
