+++ src/components/exportable/CardExportable.tsx
@@ -1,20 +1,13 @@
 import * as React from 'react';
-import { ApiClient } from '../../api/ApiClient';
 import QueryParams from '../../api/QueryParams';
 import { ISerie } from '../../api/Serie';
 import SerieApi from '../../api/SerieApi';
+import { ICardExportableConfig } from '../../indexCard';
 import FullCard from '../exportable_card/FullCard';
 
 
-export interface ICardExportableProps {
-    serieId: string;
-    locale: string;
-    links: string;
-    color: string;
-    hasChart: string;
-    chartType: string;
-    title?: string;
-    source?: string;
+export interface ICardExportableProps extends ICardExportableConfig {
+    seriesApi: SerieApi;
 }
 
 interface ICardExportableState {
@@ -22,12 +15,10 @@ interface ICardExportableState {
 }
 
 export default class CardExportable extends React.Component<ICardExportableProps, ICardExportableState> {
-    private seriesApi: SerieApi;
 
     public constructor(props: any) {
         super(props);
 
-        this.seriesApi = new SerieApi(new ApiClient(buildURLfromSerieId(props.serieId), 'ts-components-card'));
         this.state = {
             serie: null,
         }
@@ -49,13 +40,8 @@ export default class CardExportable extends React.Component<ICardExportableProps
     }
 
     private fetchSeries(params: QueryParams) {
-        this.seriesApi.fetchSeries(params)
+        this.props.seriesApi.fetchSeries(params)
             .then((series: ISerie[]) => this.setState({serie: series[0]}))
             .catch((error: any) => alert("Ocurrió un error buscando la serie solicitada."));
     }
-}
-
-
-function buildURLfromSerieId(serieId: string): string {
-    return `https://apis.datos.gob.ar/series/api/series?ids=${serieId}`
 }
